restore remote tests
