+++ tests/phpunit/src/Commands/Remote/DrushCommandTest.php
@@ -2,54 +2,63 @@
 
 namespace Acquia\Cli\Tests\Commands\Remote;
 
-use Acquia\Cli\Command\Remote\SshCommand;
+use Acquia\Cli\Command\Remote\DrushCommand;
+use Acquia\Cli\Helpers\SshHelper;
 use Prophecy\Argument;
 use Symfony\Component\Console\Command\Command;
 
 /**
- * Class SshCommandTest.
+ * Class DrushCommandTest.
  *
- * @property SshCommand $command
+ * @property DrushCommand $command
  * @package Acquia\Cli\Tests\Remote
  */
-class SshCommandTest extends SshCommandTestBase {
+class DrushCommandTest extends SshCommandTestBase {
 
   /**
    * {@inheritdoc}
    */
   protected function createCommand(): Command {
-    return $this->injectCommand(SshCommand::class);
+    return $this->injectCommand(DrushCommand::class);
   }
 
   /**
-   * Tests the 'remote:ssh' commands.
+   * Tests the 'remote:drush' commands.
    * @throws \Psr\Cache\InvalidArgumentException
+   * @throws \Exception
    */
-  public function testRemoteAliasesDownloadCommand(): void {
+  public function testRemoteDrushCommand(): void {
 
     $this->mockForGetEnvironmentFromAliasArg();
     [$process, $local_machine_helper] = $this->mockForExecuteCommand();
+
     $ssh_command = [
       'ssh',
       'site.dev@server-123.hosted.hosting.acquia.com',
       '-o StrictHostKeyChecking=no',
       '-o AddressFamily inet',
       '-o LogLevel=ERROR',
+      'cd /var/www/html/devcloud2.dev/docroot; ',
+      'drush',
+      'status',
     ];
     $local_machine_helper
       ->execute($ssh_command, Argument::type('callable'), NULL, TRUE)
       ->willReturn($process->reveal())
       ->shouldBeCalled();
-    $this->localMachineHelper = $local_machine_helper->reveal();
+    $this->command->localMachineHelper = $local_machine_helper->reveal();
+    $this->command->sshHelper = new SshHelper($this->output, $local_machine_helper->reveal());
 
     $args = [
       'alias' => 'devcloud2.dev',
+      'drush_command' => 'status',
+      '-vvv' => '',
     ];
     $this->executeCommand($args);
 
     // Assert.
     $this->prophet->checkPredictions();
-    $output = $this->getDisplay();
+    $this->getDisplay();
   }
 
 }
