+++ tests/phpunit/src/Commands/Remote/SshCommandTestBase.php
